tests/cloudfunctions/requirements.txt
functions-framework==3.5.0
google-cloud-firestore==2.13.0
google-cloud-pubsub==2.18.4

